+++ CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-// Add your changes here and then delete this line
+### Added
+
+ - Support to search multiple markets at once.
+ - Support to search all available Spotify markets.
 
 ## [2.13.0] - 2020-06-25